==> notbooks/01_preprocessing.ipynb <==
## Imports

# Imports
import pandas as pd
from collections import defaultdict
import re
from langdetect import detect
from langdetect.lang_detect_exception import LangDetectException

## Dataset Overview

# Loading the Twitter dataset
df = pd.read_csv('/Users/diegolemos/Masters/Theses/code/data/raw/twcs.csv')

# Displaying basic informations from the dataset
print(df.shape)
df.head()

# Checking for null values
df.isnull().sum()

### Null Values
As we can see above, there are a significative amount of 'response_tweet_id' and 'in_response_to_tweet_id', these missing values indicates tweets that are no part of a reply chain. For now we are going to keep them, and use these columns to recostruct conversations threads. Instead of removing or imputing them, we going to these nulls as anchor points to reconstruct complete conversation chains, by using backward traversal. This will enable us to create a new thread_id column that groups all related tweets together, which is essential for tracking customer sentiment progression across interactions.

## Conversation Thread Reconstruction

# Reconstructing conversation threads

# Creating a luokup for parent tweet IDs
id_to_parent = dict(zip(df['tweet_id'], df['in_response_to_tweet_id']))

# Tracing the root tweet ( Finding the start od the conversation)
def find_root_tweet(tweet_id, id_to_parent_cache):
    while pd.notnull(id_to_parent_cache.get(tweet_id)):
        tweet_id = id_to_parent_cache[tweet_id]
    return tweet_id

# Applting to all tweets
thread_ids = []
for tweet_id in df['tweet_id']:
    thread_id = find_root_tweet(tweet_id, id_to_parent)
    thread_ids.append(thread_id)

df['thread_id'] = thread_ids

### Reconstructing conversation threads
To understand how the messages are connected into conversation, I utilised thread_id which groups all the tweets belonging to the same conversation, with the in_response_to_tweet_id column, I was able to follow the responses back to the original tweet of a thread. This is the relation using which we will backtrack to the genesis of every tweet. We use this ID as thread_id and this way we are able to group messages belong to same support interaction.

# Checking the number of unique threads
print('Unique threads: ', df['thread_id'].nunique())

## Customer Message Filtering

# Checking amount of messages from customer (True) and agents (False)
df['inbound'].value_counts()

# Filtering customer messages
customer_df = df[df['inbound'] == True].copy()

# Dropping rows with no text
customer_df.dropna(subset = ['text'], inplace = True)
customer_df = customer_df[customer_df['text'].str.strip() != '']

# Displaying samples
customer_df['text'].sample(5).tolist()


### Customers Inbound
Here we have created a copy from the dataset containing only constumer messages as the objective of this project is to detect customer dissarisfaction.

## Text Cleaning

# Creating a function to clean the text
def clean_text(text):
    # Removing URLs
    text = re.sub(r'http\S+|www.\S+', '', text)
    
    # Removing mentions and hastags
    text = re.sub(r'@\w+|#\w+', '', text)
    
    # Removing extra space
    text = re.sub(r'\s+', ' ', text).strip()
    
    return text

# Creating a function to detect language
def detect_language(text):
    try:
        return detect(text)
    except LangDetectException:
        return 'unknown'

# Cleaning customer messagaes
customer_df['clean_text'] = customer_df['text'].apply(clean_text)

## Language Detection & Filtering

# Detecting lamguage
customer_df['lang'] = customer_df['clean_text'].apply(detect_language)

# Filtering only  for English messages
customer_df = customer_df[customer_df['lang'] == 'en'].copy()

# Printting sample with cleanned messages
customer_df[['text', 'clean_text', 'author_id']].sample(5)

# Saving the proessed dataset
customer_df.to_csv('/Users/diegolemos/Masters/Theses/code/data/processed/customer_english.csv', index=False)

### Clean text and language detection
In this step, we have cleaned the raw tweet texts by removing unnecessary elements such as URLs, mentions, hashtags nd extra space. Numbers, punctuation and emojis were kept as it can have important impact on the sentiment impressed in a message. We also have created a new column called 'clean_text' that recieves this clean text. This ensures that the data is ready for tokenisation and sentiment analysis.

Additionally, we have implemented a function to detect the language of the message, further, filtering for Eglish leaguage only, as models like VADER and SistilBERT are very English-centic.
